+++ CHANGES.md
@@ -36,6 +36,7 @@ Fixed:
 
 - Fix error when loading script path having non-ascii characters in them
   (#4343)
+- Prevent concurrent reload and request fetch in playlists (#4316)
 - Don't mark source as ready until their clock has started. (#4496)
 - Fixed mutex deadlock caused by aggressive inlining (#4540)
 - Fixed segfault when using SRT on windows (#4538)